Add recognition via sharing bananas.
You can now mention users and bananas in a message, and banabot will
track that information. There is a leaderboard that tracks the users in
the system and the bananas they are earned so far.

We don't track how many bananas a user can give, yet. The leaderboard is
only initialized once, so it likely doesn't handle the case where a user
is created after the board is created. It's probably better to not build
the board the way I am.